Host chatbot at a per-client URL
